GitOrigin-RevId: 32f0508375c4ddcc929984f84708a41e99ca4e95

+++ apps/system/components/kata-os-common/src/camkes/src/lib.rs
@@ -83,7 +83,7 @@ impl Camkes {
 
     pub fn init_allocator(self: &Camkes, heap: &'static mut [u8]) {
         unsafe {
-            allocator::ALLOCATOR.init(heap.as_mut_ptr() as usize, heap.len());
+            allocator::ALLOCATOR.init(heap.as_mut_ptr(), heap.len());
         }
         trace!("setup heap: start_addr {:p} size {}", heap.as_ptr(), heap.len());
     }
